moved one command to a commands subfolder and it worked. will start moving all commands. this is non-recursive approach to automake, as stated in https://autotools.io/automake/nonrecursive.html
